evodie

+++ src/main/java/com/example/demo/Doctor/DoctorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface DoctorRepository extends JpaRepository<Doctor, Integer> {
+public interface DoctorRepository extends JpaRepository<Doctor, Long> {
     List<Doctor>  findDoctorByClinicAndSpeciality(Integer clinicId, String speciality);
 }
